support for overwriting prototypes through plugins
