+++ plugins/typingIndicator.js
@@ -1,0 +1,11 @@
+module.exports = {
+    Chat: {
+        doStuff: function() {
+            
+            this.emitter.on('message', function(message) {
+                console.log('plugin got message', message);
+            });
+            
+        }   
+    }
+}
